"Improved data analysis and added configuration files"
In this commit, enhanced the data analysis by altering the R scripts for histograms, median calculation, box plots, and scatterplots. The changes provide improved statistics and visual representations by changing the bins for histograms, adding frequency calculations for distributions and including more descriptive labels.

The configuration files were added for Python and Version control systems. This will make sure the team works on the same software setting and sends any changes in the code directory to the version control system (Git in this case). Modifying the .gitignore file to specify files that should be avoided by the version control system helps to ensure that the RCs do not clutter with non-scarce configuration or backup files.