Refactor random sample generation and variable assignment syntax
The random sample generation command in `R_Uebung_4.1_Würfel.R` was refactored to simplify and improve readability. Redundant use of `c()` function was removed as `sample()` function can directly work with a sequence of numbers. In `R_Uebung_2.1_Hist.R`, the variable assignment operator was changed from `=` to `<-` to ensure consistency and follow the recommended R style guide.