Add new script for exercise 6.1
Added a new R script for exercise 6.1. The script runs a simulation to determine the average profit over 100,000 games where a die is thrown three times, profits are lost if a certain number (a) is not rolled, and gained if it is rolled. Results are plotted to visualise average gain over the number of runs.